+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [UNRELEASED] [3.9]
 - Added: Disable Program Compatibility Assistant Service
+- Added: Check for new versions in Options
 
 ## [3.8] - 2017-12-31
 - Windows 7 crashing fixed